#12 increase duplicated test code
create `mock` modules for the agents that are used in tests
