+++ spec/integration/full.spec.js
@@ -1,72 +1,26 @@
 /* Integration test for simplified belief-plan approach */
 const {
-  Belief,
-  Desire,
-  Intentions, // eslint-disable-line no-unused-vars
-  Plan,
   Agent,
   Environment
 } = require('../../src/js-son')
 
-describe('Environment / run()', () => {
-  const createAgent = (type) => {
-    const beliefs = {
-      ...Belief('dogNice', true),
-      ...Belief('dogHungry', false)
-    }
-    if (type === 'human') {
-      const desires = {
-        ...Desire('praiseDog', beliefs => beliefs.dogNice),
-        ...Desire('feedDog', beliefs => beliefs.dogNice && beliefs.dogHungry)
-      }
-      const preferenceFunctionGen = (beliefs, desires) => desireKey => {
-        if (!desires[desireKey](beliefs)) {
-          return false
-        } else if (desireKey === 'feedDog' || !desires['feedDog'](beliefs)) {
-          return true
-        } else {
-          return false
-        }
-      }
-      const plans = [
-        Plan(intentions => intentions.praiseDog, () => ({
-          actions: ['Good dog!']
-        })),
-        Plan(intentions => intentions.feedDog, () => ({
-          actions: ['Here, take some food!']
-        }))
-      ]
-      return new Agent('human', beliefs, desires, plans, preferenceFunctionGen)
-    } else {
-      beliefs.foodAvailable = false
-      beliefs.dogRecentlyPraised = false
-      const desires = {
-        ...Desire('wagTail', beliefs => beliefs.recentlyPraised),
-        ...Desire('eat', beliefs => beliefs.foodAvailable && beliefs.dogHungry)
-      }
-      const preferenceFunctionGen = (beliefs, desires) => desireKey => {
-        if (!desires[desireKey](beliefs)) {
-          return false
-        } else if (desireKey === 'eat' || !desires['eat'](beliefs)) {
-          return true
-        } else {
-          return false
-        }
-      }
-      const plans = [
-        Plan(intentions => intentions.eat, () => ({
-          actions: ['Eat']
-        })),
-        Plan(intentions => intentions.wagTail, () => ({
-          actions: ['Wag tail']
-        }))
-      ]
-      return new Agent('dog', beliefs, desires, plans, preferenceFunctionGen)
-    }
-  }
+const {
+  beliefs,
+  desires,
+  preferenceFunctionGen,
+  plans
+} = require('../mocks/human')
 
-  const human = createAgent('human')
-  const dog = createAgent('dog')
+const {
+  dogBeliefs,
+  dogDesires,
+  dogPreferenceFunctionGen,
+  dogPlans
+} = require('../mocks/dog')
+
+describe('Environment / run()', () => {
+  const human = new Agent('human', beliefs, desires, plans, preferenceFunctionGen)
+  const dog = new Agent('dog', dogBeliefs, dogDesires, dogPlans, dogPreferenceFunctionGen)
 
   const state = {
     dogNice: true,
