+++ spec/src/agent/Intentions.spec.js
@@ -1,13 +1,12 @@
 const Belief = require('../../../src/agent/Belief')
-const Desire = require('../../../src/agent/Desire')
 const Intentions = require('../../../src/agent/Intentions')
+
+const {
+  desires
+} = require('../../mocks/human')
 
 describe('Intentions()', () => {
   const beliefs = { ...Belief('dogNice', true), ...Belief('dogHungry', true) }
-  const desires = {
-    ...Desire('playWithDog', beliefs => beliefs.dogNice),
-    ...Desire('feedDog', beliefs => beliefs.dogNice && beliefs.dogHungry)
-  }
 
   it('should filter intentions according to provided preference function', () => {
     const preferenceFunctionGen = (beliefs, desires) => (desireKey) => {
